fix(T-28): change tenant_id column from INTEGER to BIGINT to match tenants.id type

+++ MyraTMS/scripts/058-t28-customer-os-onboarding.sql
@@ -17,7 +17,7 @@ BEGIN;
 
 CREATE TABLE IF NOT EXISTS tenant_onboarding_sessions (
     id                 SERIAL PRIMARY KEY,
-    tenant_id          INTEGER REFERENCES tenants(id),
+    tenant_id          BIGINT REFERENCES tenants(id),
 
     current_step       VARCHAR(30) NOT NULL DEFAULT 'sign_up'
                         CHECK (current_step IN (
